[#4] Fix for when cache goes bad
fixes #4

When checking the configured probes, we cache the results in a file cache.

Occassionally this cache goes bad for some reason, and Laravel returns `false` when fetching the cached object.

This change checks if the returned value is falsy, and if so returns an empty results object to start afresh.

Yes, you lose the history of your checks, but history is a secondary concern.